backend/requirements.txt
flask
pymysql

mysql-connector